// core/web-app/src/types/bindings/cloudwatch-data-metric.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface CloudwatchDataMetricMetadata { polling_interval: number, access_key: string, secret_key: string, region: string, expression: string, period: number, duration_seconds: number, }